Improve Queue, add filenames, show upload status
The queue is much more robust now, can handle failed file uploads and
lots of files. The upload queue is consumed serially just for
simplicity's sake so the user can see files consumed and the appropriate
notification after its sucessful upload. This can be improved to launch
multiple uploads in the future.

Filenames are now in a readable format i.e. month-day-year and time in
hour:minute:second:millisecond.

When a file is being uploaded its filename in the table view is
temporarliy switched to "uploading..."