Add settings with encryption key and cloud choices
Now the user can specify whether they want to upload their images to the
auditor's ipfs node or not. If they do upload the images, they can
specify if they want their local copies deleted. TODO: add if the local
copy was deleted and only the cloud version exists in the image
metadata.

The user can also specify their encryption key that they can enter in on
the auditor side to decrypt the blockchain data. It has a default
parameter set in the code.

Also added an investigation id, something that would apply to more
photographs vs the purpose of each photograph, which we assume would
vary more often.

Also changed the TWCameraView to start up the camerasession upon
receiving permission from the user for camera access.