Add unpinning from IPFS if set in settings
As a stopgap measure before we're running our own local node, if the
user sets that they don't want their photos stored in the cloud,
immediately after uploading the file and getting the CID, we unpin the
file. To be determined how long it takes after uploading the file for it
to disappear.