Add support for storage location in atra table
Now we denominate where an image is stored on the blockchain table, with
either "IPFS", "DEVICE", or "IPFS, DEVICE" options. Also made sure the
device is portrait only.

Possible TODO: when the user opens the app for the first time, they can
set a department, secret key, and username that they recall every time
they open the app and use the fingerID. If they want to sign out, they
go through this signup page again and tie it to their finrgerprint.